+++ src/config.js
@@ -42,7 +42,7 @@ export const configSyncDefaults = {
 
 export const shufflingHints = [
 	// General extension hints
-	"The extension adds a 'Shuffle' button to all channel and video pages on YouTube, this button has the same behaviour as shuffling from the popup!",
+	"The extension adds a 'Shuffle' button to all channel, video and shorts pages on YouTube, this button has the same behaviour as shuffling from the popup!",
 	"If you are the first person to shuffle from a channel, the video ID's of that channel are saved both locally and in a remote database for other users to use!",
 	"Try using the extension on April 1st - maybe something unexpected will happen!",
 	"The extension does not collect any personal information, it only stores video ID's of channels you shuffle from, without linking them back to you!",
@@ -77,7 +77,7 @@ export const shufflingHints = [
 
 	// Changelog
 	"The popup will highlight the 'Changelog' button whenever a new version of the extension has been installed!",
-	"Are you wondering how the extension has changed over time? Check out the changelog using the button on the popup!",
+	"Want to stay up-to-date on the newest features and improvements? Check out the changelog using the button in the popup!",
 
 	// Meta/GitHub
 	"Do you have an idea on how this extension could be improved? Open an issue on GitHub and let me know!",