+++ test/testSetup.js
@@ -135,7 +135,7 @@ afterEach(async function () {
 // Reimplementation of the function in the background script
 async function getAPIKey(forceDefault, useAPIKeyAtIndex = null) {
 	// List of API keys that are stored in the database/locally
-	let availableAPIKeys = null;
+	let availableAPIKeys;
 
 	// If the user has opted to use a custom API key, use that instead of the default one
 	if (!forceDefault && mockedConfigSync.useCustomApiKeyOption && mockedConfigSync.customYoutubeApiKey) {
@@ -153,8 +153,8 @@ async function getAPIKey(forceDefault, useAPIKeyAtIndex = null) {
 		return { APIKey: availableAPIKeys, isCustomKey: false, keyIndex: null };
 	}
 
-	let usedIndex = null;
-	let chosenAPIKey = null;
+	let usedIndex;
+	let chosenAPIKey;
 	if (useAPIKeyAtIndex === null) {
 		// Choose a random one of the available API keys to evenly distribute the quotas
 		usedIndex = Math.floor(Math.random() * availableAPIKeys.length);