+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v2.4.0 (Unreleased)
 
 <!--Releasenotes start-->
+- Shorts pages are now supported! Shuffle buttons can now be found on all shorts pages.
 - Added a new option to shuffle only from shorts.
 - When watching shorts, the extension will now update the last visited channel in the popup.
 - Reduced the time it takes to shuffle, as the extension now uses a more sophisticated way to decide whether or not to check if a video has been deleted or not.